Rename auth-http.ts to auth-http-helpers.ts
