+++ src/app/app.module.ts
@@ -20,7 +20,7 @@ import { IonicPlugins } from '../providers/ionic-plugins';
 import { UserData } from '../providers/user-data';
 
 import { RavenErrorHandler } from '../services/raven-error-handler';
-import { getAuthHttp, cloudSettings } from '../services/auth-http';
+import { getAuthHttp, cloudSettings } from '../services/auth-http-helpers';
 
 @NgModule({
   declarations: [
